rename header file for temp_sensor.h
